GATT: UnixStream -> UnixDatagram (#172)
GATT communication with bluetoothd happens over a SOCK_SEQPACKET socket, however bluer uses UnixStream socket for those sockets, this leads to issues with tokio/mio/epoll, when receiving notification data from from a remote GATT characteristic. The problem works as follows:

    - the bluer application does initialization and subscribes to GATT notification events, tokio/mio puts the UnixStream socket into epoll using edge trigger mode.
    - Remote service send multiple notification event(at least 3?)
    - The bluer application wakes from epoll and receives the first event, it then begins processing the first event.
    - While the application processes the event, more events are received(at least 2 more)
    - The application finishes processing of the event and call epoll_wait, as there has been more new events since last epoll the application is woken immediately to process the second event.
    - When the process enter epoll_wait this time it will not get woken again, even if there are still events to process, as there are no NEW events since last epoll. The application will only receive the next event when an additional new event is received, however this new event will remain in the queue and not be processed.

This happens because tokio/mio assumes that on a UnixStream socket there is no more data available if the read does not fill the whole buffer. However the correct thing to do on a SOCK_SEQPACKET socket is to read from the socket until it returns EAGAIN processing every packet in the queue.

This is what UnixDatagram socket does.

This commit uses UnixDatagram sockets instead of UnixStream sockets for GATT Characteristic communication.